+++ TinkoffASDKUI/TinkoffASDKUI/AppBasedTDSFlow/TDSController.swift
@@ -200,8 +200,9 @@ extension TDSController: ChallengeStatusReceiver {
         finishTransaction()
         do {
             let cresValue = try buildCresValue(with: completionEvent.getTransactionStatus())
+            let data = Submit3DSAuthorizationV2Data(cres: cresValue, paymentId: nil)
 
-            acquiringSdk.submit3DSAuthorizationV2(cres: cresValue) { [weak self] result in
+            acquiringSdk.submit3DSAuthorizationV2(data: data) { [weak self] result in
                 self?.completionHandler?(result)
                 self?.clear()
             }